Move some commands to the Hunger client system so they'll be available to clients, not just a host player.
