+++ src/main/java/org/terasology/hunger/HungerAuthoritySystem.java
@@ -159,87 +159,4 @@ public class HungerAuthoritySystem extends BaseComponentSystem implements Update
             inventoryManager.removeItem(event.getInstigator(), event.getInstigator(), slot, true, 1);
         }
     }
-
-    /**
-     * Tests the character ID and content of all components for an Entity.
-     * @param client
-     */
-    @Command(shortDescription = "Temp testing command", runOnServer = true)
-    public void test(EntityRef client) {
-        logger.info("ID:" + client.getComponent(ClientComponent.class).character.getId());
-        for (Component c : client.getComponent(ClientComponent.class).character.iterateComponents()) {
-            logger.info(c.toString());
-        }
-    }
-
-    /**
-     * A command for testing the hunger level for an entity.
-     * @param client The entity who is checking it's hunger level.
-     * @return Returns a message for the client informing them about their food level if they have one.
-     */
-    @Command(shortDescription = "Checks your hunger/food level.", runOnServer = true)
-    public String hungerCheck(@Sender EntityRef client) {
-        EntityRef character = client.getComponent(ClientComponent.class).character;
-        if (character.hasComponent(HungerComponent.class)) {
-            HungerComponent hunger = character.getComponent(HungerComponent.class);
-            return "Current Food Level: " + HungerUtils.getHungerForEntity(character) + "/" + hunger.maxFoodCapacity;
-        } else {
-            return "You don't have a hunger level.";
-        }
-    }
-
-    /**
-     * A command for modifying your hunger level.
-     * @param newFood The new hunger level for the client. This has to be above 0 and below the max food capacity.
-     * @param client The client which is changing it's hunger level.
-     * @return Returns a message for the client telling him about their new hunger level if they have one.
-     */
-    @Command(shortDescription = "Sets your current hunger level.", runOnServer = true)
-    public String hungerSet(@CommandParam(value = "FoodLevel") float newFood, @Sender EntityRef client) {
-        EntityRef character = client.getComponent(ClientComponent.class).character;
-        if (!character.hasComponent(HungerComponent.class)) {
-            return "You don't have a hunger level.";
-        }
-        HungerComponent hunger = character.getComponent(HungerComponent.class);
-        if (newFood < 0) {
-            hunger.lastCalculatedFood = 0;
-            hunger.lastCalculationTime = time.getGameTimeInMs();
-            character.saveComponent(hunger);
-            return "Food level cannot be below 0. Setting to 0.";
-        }
-        if (newFood > hunger.maxFoodCapacity) {
-            hunger.lastCalculatedFood = hunger.maxFoodCapacity;
-            hunger.lastCalculationTime = time.getGameTimeInMs();
-            character.saveComponent(hunger);
-            return "Food level cannot be above Max Food Capacity. Setting to Max(" + hunger.maxFoodCapacity + ")";
-        }
-        hunger.lastCalculatedFood = newFood;
-        hunger.lastCalculationTime = time.getGameTimeInMs();
-        character.saveComponent(hunger);
-        return "Food level successfully set to: " + newFood;
-    }
-
-    /**
-     * A command for changing your maximum food level.
-     * @param newMax The new maximum food level. Has to be above 0.
-     * @param client The client which is changing it's food level.
-     * @return Returns a message for the client telling him wether the command was succesful.
-     */
-    @Command(shortDescription = "Sets your max food level.", runOnServer = true)
-    public String hungerSetMax(@CommandParam(value = "MaxFoodLevel") float newMax, @Sender EntityRef client) {
-        EntityRef character = client.getComponent(ClientComponent.class).character;
-        if (!character.hasComponent(HungerComponent.class)) {
-            return "You don't have a hunger level.";
-        }
-        HungerComponent hunger = character.getComponent(HungerComponent.class);
-        if (newMax <= 0) {
-            hunger.maxFoodCapacity = 100;
-            character.saveComponent(hunger);
-            return "Max Food Level cannot be below or equal to 0. Setting to default (100)";
-        }
-        hunger.maxFoodCapacity = newMax;
-        character.saveComponent(hunger);
-        return "Max Food Level successfully set to: " + newMax;
-    }
-
 }
